fix(rename-overwrite): limit the amount of retries

+++ rename-overwrite/index.js
@@ -3,16 +3,18 @@ const fs = require('fs')
 const path = require('path')
 const rimraf = require('@zkochan/rimraf')
 
-module.exports = async function renameOverwrite (oldPath, newPath) {
+module.exports = async function renameOverwrite (oldPath, newPath, retry = 0) {
   try {
     await fs.promises.rename(oldPath, newPath)
   } catch (err) {
+    retry++
+    if (retry > 3) throw err
     switch (err.code) {
       case 'ENOTEMPTY':
       case 'EEXIST':
       case 'ENOTDIR':
         await rimraf(newPath)
-        await renameOverwrite(oldPath, newPath)
+        await renameOverwrite(oldPath, newPath, retry)
         break
       // Windows Antivirus issues
       case 'EPERM':
@@ -45,7 +47,7 @@ module.exports = async function renameOverwrite (oldPath, newPath) {
           }
         }
         await fs.promises.mkdir(path.dirname(newPath), { recursive: true })
-        await renameOverwrite(oldPath, newPath)
+        await renameOverwrite(oldPath, newPath, retry)
         break
       default:
         throw err
@@ -53,10 +55,12 @@ module.exports = async function renameOverwrite (oldPath, newPath) {
   }
 }
 
-module.exports.sync = function renameOverwriteSync (oldPath, newPath) {
+module.exports.sync = function renameOverwriteSync (oldPath, newPath, retry = 0) {
   try {
     fs.renameSync(oldPath, newPath)
   } catch (err) {
+    retry++
+    if (retry > 3) throw err
     switch (err.code) {
       case 'ENOTEMPTY':
       case 'EEXIST':
@@ -67,7 +71,7 @@ module.exports.sync = function renameOverwriteSync (oldPath, newPath) {
         return
       case 'ENOENT':
         fs.mkdirSync(path.dirname(newPath), { recursive: true })
-        renameOverwriteSync(oldPath, newPath)
+        renameOverwriteSync(oldPath, newPath, retry)
         return
       default:
         throw err
